requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
